updating front app

+++ todo-app-front/src/ui/component/Task/Task.component.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { TaskApi } from '../../../hooks/task/task.hook'
 import './style.css'
 
-export function Task({task, update}) {
+export function Task({ task, update }) {
 
     const taskApi = TaskApi()
 
@@ -11,38 +11,43 @@ export function Task({task, update}) {
         return string.charAt(0).toUpperCase() + string.slice(1)
     }
 
-    async function handleDeleteTask() { 
+    async function handleDeleteTask() {
         try {
-            const response = await taskApi.deleteTask(task.id)
-            update()
-        } catch(err) {
-            console.log(err)
+            if (window.confirm("Are you sure?")) {
+                await taskApi.deleteTask(task.id)
+                update()
+            }
+        } catch (err) {
+            window.alert("An unexpected error occurred. Try again later!")
         }
     }
 
     async function handleFinishTask() {
         try {
-            const response = await taskApi.finishTask(task.id)
+            await taskApi.finishTask(task.id)
             update()
-        } catch(err) {
-            console.log(err)
+        } catch (err) {
+            window.alert("An unexpected error occurred. Try again later!")
         }
     }
-
 
     return (
         <div className="task-container">
             <header>
-                <span className={task.priority == 'URGENT' ? 'urgent' : undefined}>
+                <h2 className={task.priority === 'URGENT' ? 'urgent' : undefined}>
                     {capitalizeFirstLetter(task.priority)}
-                </span>
+                </h2>
             </header>
             <main>
-                <p className={task.hasFinished ? 'line-through' : undefined}><strong>Description: </strong>{task.description}</p>
-                <p className={task.hasFinished ? 'line-through' : undefined}><strong>Prevision: </strong>{`${task.conclusionPrevision[0]}-${task.conclusionPrevision[1]}-${task.conclusionPrevision[2]}`}</p>
+                <p className={task.hasFinished ? 'line-through' : undefined}>
+                    <strong>Description: </strong>{task.description}
+                </p>
+                <p className={task.hasFinished ? 'line-through' : undefined}>
+                    <strong>Prevision: </strong>{`${task.conclusionPrevision[0]}/${task.conclusionPrevision[1]}/${task.conclusionPrevision[2]}`}
+                </p>
             </main>
             <footer>
-                {!task.hasFinished && <button className="check-img" onClick={handleFinishTask}></button> }
+                {!task.hasFinished && <button className="check-img" onClick={handleFinishTask}></button>}
                 <button className="remove-img" onClick={handleDeleteTask}></button>
             </footer>
         </div>
